Renaming things a bit and switching to dynamic package description.
