--- setup.cfg
@@ -1,24 +1,0 @@
-[metadata]
-name = differential-value-iteration
-version = 0.0.1
-author = Abhishek Naik
-author_email = abhisheknaik22296@gmail.com
-description = Experimental algorithms for differential value iteration.
-long_description = file: README.md
-long_description_content_type = text/markdown
-url = https://github.com/abhisheknaik96/differential-value-iteration/
-project_urls =
-    Bug Tracker = ttps://github.com/abhisheknaik96/differential-value-iteration/issues
-classifiers =
-    Programming Language :: Python :: 3
-    License :: OSI Approved :: MIT License
-    Operating System :: OS Independent
-
-[options]
-package_dir =
-    = src
-packages = find:
-python_requires = >=3.6
-
-[options.packages.find]
-where = src
